Split dnd components to deparate modules.
